fix/board: Collect all signals & register in AppConfig.
